Creating scripts directory and adding netdump.sh
Networking assignment - Write a script to dump network info (including ip, dns, open ports and anything else you think is useful) to the command line and to a file.  Make sure the name of the file includes the date and the purpose.  This needs to be more then just a single command or two, really think about the format of the file it's going into, what's useful, and what you would want if someone handed you this file.

Uploaded netdump.sh, a shell script I created that will check or a folder called /logs/netdump and create it if it doesn't exist. It will then check to see if the host machine is Centos or Ubuntu and carry out it's commands to create a log file with the current day and time, display and record the user executing the script, host machine name, host machine IP address, default gateway and subnet mask, check status of fail2ban, check firewall status, show ports with active traffic,  and run a scan of all open and filtered ports using nmap. All results are then printed to the new log file located in /logs/netdump/