Added original netdump version
This version I had originally done to just get the script running in Centos. The signs of it starting to change over to trying to work on both OSes you can see with the if/else statement up at the top where it checks the $osflavor. From there it grew into having to have different statements on most of the commands inside the script based on the OS you were using and if the repo was installed or not. I did consider changing the name because of the service with the same name but for just class purposes I'm going to keep it.